# Conflicts:
#	app/Http/Controllers/Admin/UserController.php

+++ database/migrations/2023_03_16_061811_create_pre_order_vshop_table.php
@@ -33,7 +33,7 @@ return new class extends Migration
             $table->string("address");
             $table->string("no");
 
-            $table->decimal("total", 11, 3);
+            $table->double("total", 20, 2);
             $table->decimal("discount", 11, 3);
             $table->decimal("deposit_money", 11, 3);
